creacion de helper en view y formateo html
