+++ app/src/main/java/ayds/newyork/songinfo/moredetails/model/repository/external/infos/NYTimesToInfoResolver.kt
@@ -13,9 +13,6 @@ private const val RESPONSE = "response"
 private const val DOCS = "docs"
 private const val ABSTRACT = "abstract"
 private const val WEB_URL = "web_url"
-private const val HTML_DIV_WIDTH = "<html><div width=400>"
-private const val HTML_FONT = "<font face=\"arial\">"
-private const val HTML_END_TAGS = "</font></div></html>"
 
 internal class JsonToInfoResolver : NYTimesToInfoResolver {
 
@@ -40,30 +37,11 @@ internal class JsonToInfoResolver : NYTimesToInfoResolver {
     }
 
     private fun JsonObject.getArtistInformation(artistName: String): String {
-        return textToHtml(getCleanAbstract(), artistName)
-    }
-
-    private fun JsonObject.getCleanAbstract(): String {
         val abstract = this[DOCS].asJsonArray[0].asJsonObject[ABSTRACT]
         return abstract.asString.replace("\\n", "\n")
     }
 
     private fun JsonObject.getUrl(): String {
         return this[DOCS].asJsonArray[0].asJsonObject[WEB_URL].asString
-    }
-
-    private fun textToHtml(text: String, termToBold: String): String {
-        return StringBuilder().apply {
-            append(HTML_DIV_WIDTH)
-            append(HTML_FONT)
-            append(formatTextToHtmlWithBoldTerm(text, termToBold))
-            append(HTML_END_TAGS)
-        }.toString()
-    }
-
-    private fun formatTextToHtmlWithBoldTerm(text: String, termToBold: String): String {
-        return text.replace("'", " ")
-            .replace("\n", "<br>")
-            .replace(termToBold.toRegex(), "<b>" + termToBold.uppercase() + "</b>")
     }
 }
